requirements.txt

streamlit==1.25.0
langchain>=0.0.300
openai>=0.27.0
faiss-cpu>=1.7.4
sentence-transformers>=2.2.2
transformers>=4.30.0
python-dotenv>=1.0.0
langgraph>=0.1.0  # placeholder: install actual LangGraph package if named differently
typing_extensions
